reinventing lerp

+++ Elaborate/Coroutine/CustomYieldInstructions/WaitForSeconds.cs
@@ -2,12 +2,12 @@ using Godot;
 
 public class WaitForSeconds : CustomYieldInstruction
 {
-    private float _time_end;
+    private double _time_end;
 
-    private float CurrentTime => Time.GetTicksMsec();
+    private double CurrentTime => Time.GetTicksMsec();
     public override bool KeepWaiting => CurrentTime < _time_end;
 
-    public WaitForSeconds(float seconds)
+    public WaitForSeconds(double seconds)
     {
         _time_end = CurrentTime + seconds * 1000;
     }
